tcp: replace tcp_time_now with tcp_time_now_us
It looks like tcp_time_now has been deprecated for a while and the
replacement is tcp_time_now_us

Type: fix

+++ src/vnet/session/transport_types.h
@@ -113,7 +113,7 @@ typedef struct _transport_connection
 
 #if TRANSPORT_DEBUG
   elog_track_t elog_track;	/**< Event logging */
-  u32 cc_stat_tstamp;		/**< CC stats timestamp */
+  f64 cc_stat_tstamp;		/**< CC stats timestamp */
 #endif
 
   /**
